requirements.txt
tensorflow-cpu
streamlit
streamlit-drawable-canvas
numpy
scikit-image
